+++ src/hypua2jamo/_cython.pyx
@@ -28,31 +28,31 @@ cdef extern from *:
     unicode PyUnicode_FromUnicode(Py_UNICODE *u, Py_ssize_t size)
 
 cdef extern from "hypua2jamo.h":
-    int hypua_c2d_ucs4_calcsize(void *src, int srclen);
-    ptrdiff_t hypua_c2d_ucs4_encode(void *src, int srclen, void *dst);
-    int hypua_c2d_ucs2_calcsize(void *src, int srclen);
-    ptrdiff_t hypua_c2d_ucs2_encode(void *src, int srclen, void *dst);
-    int hypua_p2jc_ucs4_calcsize(void *src, int srclen);
-    ptrdiff_t hypua_p2jc_ucs4_encode(void *src, int srclen, void *dst);
-    int hypua_p2jc_ucs2_calcsize(void *src, int srclen);
-    ptrdiff_t hypua_p2jc_ucs2_encode(void *src, int srclen, void *dst);
-    int hypua_p2jd_ucs4_calcsize(void *src, int srclen);
-    ptrdiff_t hypua_p2jd_ucs4_encode(void *src, int srclen, void *dst);
-    int hypua_p2jd_ucs2_calcsize(void *src, int srclen);
-    ptrdiff_t hypua_p2jd_ucs2_encode(void *src, int srclen, void *dst);
+    size_t hypua_c2d_ucs4_calcsize(void *src, size_t srclen);
+    size_t hypua_c2d_ucs4_encode(void *src, size_t srclen, void *dst);
+    size_t hypua_c2d_ucs2_calcsize(void *src, size_t srclen);
+    size_t hypua_c2d_ucs2_encode(void *src, size_t srclen, void *dst);
+    size_t hypua_p2jc_ucs4_calcsize(void *src, size_t srclen);
+    size_t hypua_p2jc_ucs4_encode(void *src, size_t srclen, void *dst);
+    size_t hypua_p2jc_ucs2_calcsize(void *src, size_t srclen);
+    size_t hypua_p2jc_ucs2_encode(void *src, size_t srclen, void *dst);
+    size_t hypua_p2jd_ucs4_calcsize(void *src, size_t srclen);
+    size_t hypua_p2jd_ucs4_encode(void *src, size_t srclen, void *dst);
+    size_t hypua_p2jd_ucs2_calcsize(void *src, size_t srclen);
+    size_t hypua_p2jd_ucs2_encode(void *src, size_t srclen, void *dst);
 
-    int hypua_d2c_ucs4_calcsize(void *src, int srclen);
-    ptrdiff_t hypua_d2c_ucs4_decode(void *src, int srclen, void *dst);
-    int hypua_d2c_ucs2_calcsize(void *src, int srclen);
-    ptrdiff_t hypua_d2c_ucs2_decode(void *src, int srclen, void *dst);
-    int hypua_jc2p_ucs4_calcsize(void *src, int srclen);
-    ptrdiff_t hypua_jc2p_ucs4_decode(void *src, int srclen, void *dst);
-    int hypua_jc2p_ucs2_calcsize(void *src, int srclen);
-    ptrdiff_t hypua_jc2p_ucs2_decode(void *src, int srclen, void *dst);
-    int hypua_jd2p_ucs4_calcsize(void *src, int srclen);
-    ptrdiff_t hypua_jd2p_ucs4_decode(void *src, int srclen, void *dst);
-    int hypua_jd2p_ucs2_calcsize(void *src, int srclen);
-    ptrdiff_t hypua_jd2p_ucs2_decode(void *src, int srclen, void *dst);
+    size_t hypua_d2c_ucs4_calcsize(void *src, size_t srclen);
+    size_t hypua_d2c_ucs4_decode(void *src, size_t srclen, void *dst);
+    size_t hypua_d2c_ucs2_calcsize(void *src, size_t srclen);
+    size_t hypua_d2c_ucs2_decode(void *src, size_t srclen, void *dst);
+    size_t hypua_jc2p_ucs4_calcsize(void *src, size_t srclen);
+    size_t hypua_jc2p_ucs4_decode(void *src, size_t srclen, void *dst);
+    size_t hypua_jc2p_ucs2_calcsize(void *src, size_t srclen);
+    size_t hypua_jc2p_ucs2_decode(void *src, size_t srclen, void *dst);
+    size_t hypua_jd2p_ucs4_calcsize(void *src, size_t srclen);
+    size_t hypua_jd2p_ucs4_decode(void *src, size_t srclen, void *dst);
+    size_t hypua_jd2p_ucs2_calcsize(void *src, size_t srclen);
+    size_t hypua_jd2p_ucs2_decode(void *src, size_t srclen, void *dst);
 
     int hypua_decoder_alloc_size();
     void hypua_decoder_init_jc2p(void *decoder);
@@ -66,23 +66,23 @@ cdef extern from "hypua2jamo.h":
     );
     int hypua_decoder_getstate(void *decoder);
     int hypua_decoder_setstate(void *decoder, int state);
-    int hypua_decoder_calcsize_ucs2(void *decoder, void *src, int srclen);
-    int hypua_decoder_calcsize_ucs4(void *decoder, void *src, int srclen);
-    int hypua_decoder_calcsize_flush(void *decoder);
-    ptrdiff_t hypua_decoder_decode_ucs2(
+    size_t hypua_decoder_calcsize_ucs2(void *decoder, void *src, size_t srclen);
+    size_t hypua_decoder_calcsize_ucs4(void *decoder, void *src, size_t srclen);
+    size_t hypua_decoder_calcsize_flush(void *decoder);
+    size_t hypua_decoder_decode_ucs2(
         void *decoder,
         void *src,
-        int srclen,
+        size_t srclen,
         void *dst
     );
-    ptrdiff_t hypua_decoder_decode_ucs4(
+    size_t hypua_decoder_decode_ucs4(
         void *decoder,
         void *src,
-        int srclen,
+        size_t srclen,
         void *dst
     );
-    ptrdiff_t hypua_decoder_decode_flush_ucs2(void *decoder, void *dst);
-    ptrdiff_t hypua_decoder_decode_flush_ucs4(void *decoder, void *dst);
+    size_t hypua_decoder_decode_flush_ucs2(void *decoder, void *dst);
+    size_t hypua_decoder_decode_flush_ucs4(void *decoder, void *dst);
 
 
 cdef int _UNICODE_SIZE = array('u').itemsize
@@ -91,10 +91,10 @@ cdef int _UNICODE_SIZE = array('u').itemsize
 cdef class JamoDecoderImplementationOnCython:
 
     cdef void* _decoder
-    cdef int (*_calcsize)(void*, void*, int)
-    cdef int (*_calcsize_flush)(void *)
-    cdef ptrdiff_t (*_decode)(void*, void*, int, void*)
-    cdef ptrdiff_t (*_decode_flush)(void*, void*)
+    cdef size_t (*_calcsize)(void*, void*, size_t)
+    cdef size_t (*_calcsize_flush)(void *)
+    cdef size_t (*_decode)(void*, void*, size_t, void*)
+    cdef size_t (*_decode_flush)(void*, void*)
 
     def __dealloc__(self):
         PyMem_Free(self._decoder)
@@ -245,8 +245,8 @@ cdef class JamoEncoderImplementationOnCython:
     Cython implementation.
     '''
 
-    cdef int (*_calcsize)(void *src, int srclen);
-    cdef ptrdiff_t (*_encode)(void *src, int srclen, void *dst);
+    cdef size_t (*_calcsize)(void *src, size_t srclen);
+    cdef size_t (*_encode)(void *src, size_t srclen, void *dst);
 
     def reset(self):
         pass
